refactor: remove barrel file for global components, store
