--- src/components/index.ts
@@ -1,2 +1,0 @@
-export * from './error-catcher'
-export * from './base/icon-link'
